requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
PyJWT
sqlalchemy
psycopg2-binary
